Refactor: Simplify ChatWidget initialization and enhance error handling
- Removed the PullseNamespace and integrated direct rendering of the ChatWidget component.
- Implemented QueryClient and TooltipProvider for improved state management and UI interactions.
- Updated error handling to provide clearer feedback during initialization failures.
- Adjusted ChatWidget to retrieve workspaceId and apiKey from local storage instead of passing them as props.
- Enhanced useWidgetConfig and useChatState hooks to utilize the updated API for fetching configuration and managing state.
